Implement interface for all custom settings
